+++ gribapi/gribapi.py
@@ -11,7 +11,8 @@
 
 """
 @package gribapi
-@brief This package is the \b Python 3 interface to ecCodes. It offers almost one to one bindings to the C API functions.
+@brief This package is the \b Python 3 interface to ecCodes.
+It offers almost one to one bindings to the C API functions.
 
 The Python 3 interface to ecCodes uses the <a href="http://numpy.scipy.org/"><b>NumPy</b></a> package
 as the container of choice for the possible arrays of values that can be encoded/decoded in and from a message.
@@ -24,14 +25,16 @@ Numpy is a package used for scientific computing in Python and an efficient cont
 
 """
 
-from functools import wraps
-import sys
 import os
+import sys
+from functools import wraps
+
 import numpy as np
 
-from .bindings import ENC, ffi, lib
-from .bindings import __version__ as bindings_version  # noqa
 from . import errors
+from .bindings import ENC
+from .bindings import __version__ as bindings_version  # noqa
+from .bindings import ffi, lib, library_path
 
 try:
     type(file)
@@ -241,7 +244,7 @@ def gts_new_from_file(fileobj, headers_only=False):
     @return               id of the GTS loaded in memory or None
     @exception CodesInternalError
     """
-    # err, h = err_last(lib.gts_new_from_file)(ffi.NULL, fileobj)
+
     err, h = err_last(lib.codes_handle_new_from_file)(
         ffi.NULL, fileobj, CODES_PRODUCT_GTS
     )
@@ -270,7 +273,7 @@ def metar_new_from_file(fileobj, headers_only=False):
     @return               id of the METAR loaded in memory or None
     @exception CodesInternalError
     """
-    # err, h = err_last(lib.metar_new_from_file)(ffi.NULL, fileobj)
+
     err, h = err_last(lib.codes_handle_new_from_file)(
         ffi.NULL, fileobj, CODES_PRODUCT_METAR
     )
@@ -397,7 +400,7 @@ def grib_new_from_file(fileobj, headers_only=False):
     @return               id of the grib loaded in memory or None
     @exception CodesInternalError
     """
-    # err, h = err_last(lib.grib_new_from_file)(ffi.NULL, fileobj, headers_only)
+
     err, h = err_last(lib.codes_handle_new_from_file)(
         ffi.NULL, fileobj, CODES_PRODUCT_GRIB
     )
@@ -1315,7 +1318,10 @@ def grib_index_new_from_file(filename, keys):
 
     @param filename   path of the file to index on
     @param keys       sequence of keys to index on.
-                      The type of the key can be explicitly declared appending :l for long (or alternatively :i), :d for double, :s for string to the key name.
+                      The type of the key can be explicitly declared appending
+                      :l for long (or alternatively :i),
+                      :d for double,
+                      :s for string to the key name.
     @return           index id
     @exception CodesInternalError
     """
@@ -1383,7 +1389,8 @@ def grib_index_get_long(indexid, key):
     @brief Get the distinct values of the key in argument contained in the index.
     The key must belong to the index.
 
-    This function is used when the type of the key was explicitly defined as long or when the native type of the key is long.
+    This function is used when the type of the key was explicitly defined as long or when the native type of
+    the key is long.
 
     \b Examples: \ref grib_index.py "grib_index.py"
 
@@ -1408,7 +1415,8 @@ def grib_index_get_string(indexid, key):
     @brief Get the distinct values of the key in argument contained in the index.
     The key must belong to the index.
 
-    This function is used when the type of the key was explicitly defined as string or when the native type of the key is string.
+    This function is used when the type of the key was explicitly defined as string or when the native type of
+    the key is string.
 
     \b Examples: \ref grib_index.py "grib_index.py"
 
@@ -1434,7 +1442,8 @@ def grib_index_get_double(indexid, key):
     @brief Get the distinct values of the key in argument contained in the index.
     The key must belong to the index.
 
-    This function is used when the type of the key was explicitly defined as double or when the native type of the key is double.
+    This function is used when the type of the key was explicitly defined as double or when the native type
+    of the key is double.
 
     \b Examples: \ref grib_index.py "grib_index.py"
 
@@ -1448,9 +1457,9 @@ def grib_index_get_double(indexid, key):
 
     values_p = ffi.new("double[]", nval)
     size_p = ffi.new("size_t *", nval)
-    err = lib.grib_index_get_doule(ih, key.encode(ENC), values_p, size_p)
+    err = lib.grib_index_get_double(ih, key.encode(ENC), values_p, size_p)
     GRIB_CHECK(err)
-    return tuple(int(values_p[i]) for i in range(size_p[0]))
+    return tuple(values_p[i] for i in range(size_p[0]))
 
 
 @require(indexid=int, key=str, value=int)
@@ -1459,7 +1468,8 @@ def grib_index_select_long(indexid, key, value):
     @brief Select the message subset with key==value.
     The value is an integer.
 
-    The key must have been created with integer type or have integer as native type if the type was not explicitly defined in the index creation.
+    The key must have been created with integer type or have integer as native type if the type
+    was not explicitly defined in the index creation.
 
     \b Examples: \ref grib_index.py "grib_index.py"
 
@@ -1478,7 +1488,8 @@ def grib_index_select_double(indexid, key, value):
     @brief Select the message subset with key==value.
     The value is a double.
 
-    The key must have been created with integer type or have integer as native type if the type was not explicitly defined in the index creation.
+    The key must have been created with integer type or have integer as native type if the type was
+    not explicitly defined in the index creation.
 
     \b Examples: \ref grib_index.py "grib_index.py"
 
@@ -1497,7 +1508,8 @@ def grib_index_select_string(indexid, key, value):
     @brief Select the message subset with key==value.
     The value is an integer.
 
-    The key must have been created with string type or have string as native type if the type was not explicitly defined in the index creation.
+    The key must have been created with string type or have string as native type if the type
+    was not explicitly defined in the index creation.
 
     \b Examples: \ref grib_index.py "grib_index.py"
 
@@ -1516,7 +1528,8 @@ def grib_new_from_index(indexid):
     @brief Create a new handle from an index after having selected the key values.
 
     All the keys belonging to the index must be selected before calling this function.
-    Successive calls to this function will return all the handles compatible with the constraints defined selecting the values of the index keys.
+    Successive calls to this function will return all the handles compatible with the constraints
+    defined selecting the values of the index keys.
 
     The message can be accessed through its gribid and will be available until @ref grib_release is called.
 
@@ -1818,7 +1831,8 @@ def grib_find_nearest(gribid, inlat, inlon, is_lsm=False, npoints=1):
 @require(gribid=int, is_lsm=bool)
 def grib_find_nearest_multiple(gribid, is_lsm, inlats, inlons):
     """
-    @brief Find the nearest point of a set of points whose latitudes and longitudes are given in the inlats, inlons arrays respectively
+    @brief Find the nearest point of a set of points whose latitudes and longitudes are given in
+    the inlats, inlons arrays respectively
 
     @param gribid     id of the GRIB message loaded in memory
     @param is_lsm     True if the nearest land point is required otherwise False.
@@ -2043,8 +2057,11 @@ def grib_set(msgid, key, value):
     #    # The value passed in is iterable; i.e. a list or array etc
     #    grib_set_array(msgid, key, value)
     else:
+        hint = ""
+        if hasattr(value, "__iter__"):
+            hint = " (Hint: for array keys use codes_set_array(msgid, key, value))"
         raise errors.GribInternalError(
-            "Invalid type of value when setting key '%s'." % key
+            "Invalid type of value when setting key '%s'%s." % (key, hint)
         )
 
 
@@ -2389,6 +2406,7 @@ def codes_bufr_extract_headers(filepath, is_strict=True):
         i += 1
 
 
+@require(msgid=int)
 def codes_bufr_key_is_header(msgid, key):
     """
     @brief Check if the BUFR key is in the header or in the data section.
@@ -2509,3 +2527,7 @@ def grib_nearest_find(nid, gribid, inlat, inlon, flags, is_lsm=False, npoints=4)
         )
 
     return tuple(result)
+
+
+def codes_get_library_path():
+    return library_path