+++ CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog for eccodes-python
 ============================
 
+
+1.3.1 (2021-04-16)
+--------------------
+
+- Fix the recommended version
+
+
+1.3.0 (2021-04-09)
+--------------------
+
+- ECC-1231: Remove the experimental high-level interface
+- Added the "findlibs" module
+- Fix tests/test_high_level_api.py when MEMFS enabled
+- ECC-1226: Python3 bindings: Typo causes AttributeError when calling codes_index_get_double
+
+
 1.2.0 (2021-03-23)
 --------------------